+++ foursouls/model/phase.py
@@ -16,9 +16,11 @@ class TurnFlags:
     loot_plays_allowed: int = 1
     attack_used: bool = False
     purchase_used: bool = False
+    died_this_turn: bool = False
 
     def reset(self) -> None:
         self.loot_plays_used = 0
         self.loot_plays_allowed = 1
         self.attack_used = False
         self.purchase_used = False
+        self.died_this_turn = False